editar user + ventana de dialogo confirmar delete

+++ FrontEnd/ampa/src/app/modules/dashboard/dashboard/clientes/clientes.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Usuario } from 'src/app/modules/productos/interfaces/usuario.interface';
 import { UsuariosService } from '../../usuarios.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from './confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-clientes',
@@ -10,11 +12,13 @@ import { UsuariosService } from '../../usuarios.service';
 })
 export class ClientesComponent implements OnInit {
   
+  selectedUserId: number | undefined;
+  editMode = false;
   usuarios: Usuario[] = [];
   usuarioForm: FormGroup;
   mostrarFormulario = false;
 
-  constructor(private usuariosService: UsuariosService, private fb: FormBuilder) {
+  constructor(private usuariosService: UsuariosService, private fb: FormBuilder, private dialog: MatDialog) {
     this.usuarioForm = this.fb.group({
       dni: ['', Validators.required],
       nombre: ['', Validators.required],
@@ -47,20 +51,55 @@ export class ClientesComponent implements OnInit {
     });
   }
 
-  editarUsuario(usuario: Usuario): void {
+  editarUsuario(usuario: Usuario, id: number): void {
+    this.editMode = true;
     this.usuarioForm.patchValue(usuario);
+    this.selectedUserId = id;
     this.mostrarFormulario = true;
   }
 
   borrarUsuario(usuario: Usuario): void {
-    this.usuariosService.deleteUser(usuario.id).subscribe(() => {
-      const index = this.usuarios.findIndex(u => u.id === usuario.id);
-      this.usuarios.splice(index, 1);
-      // Obtener la lista actualizada de usuarios
-      this.fetchData();
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        message: `Seguro quieres eliminar a ${usuario.nombre} ${usuario.apellido}? Se eliminará definitivamente.`
+      }
+    });
+  
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.usuariosService.deleteUser(usuario.id).subscribe(() => {
+          const index = this.usuarios.findIndex(u => u.id === usuario.id);
+          this.usuarios.splice(index, 1);
+          // Obtener la lista actualizada de usuarios
+          this.fetchData();
+        });
+      }
     });
   }
 
+  guardarUsuario(): void {
+    const usuario = this.usuarioForm.value;
+    this.mostrarFormulario = !this.mostrarFormulario
+    if (this.editMode) {
+      if (this.selectedUserId) {
+        usuario.id = this.selectedUserId; // Asignar el ID del usuario al objeto usuario
+        this.usuariosService.updateUser(usuario).subscribe(updatedUser => {
+          // Realizar cualquier acción adicional después de editar el usuario
+          this.editMode = false;
+          this.fetchData();
+        });
+      } else {
+        console.error("El ID del usuario es indefinido.");
+      }
+    } else {
+      this.usuariosService.addUser(usuario).subscribe(newUser => {
+        // Realizar cualquier acción adicional después de agregar el usuario
+        this.fetchData();
+      });
+    }
+  }
+
+  mostrarForm(){!this.mostrarFormulario};
   ocultarForm() {
     setTimeout(() => {
       this.mostrarFormulario = false;
